+++ test/national_id_test.dart
@@ -1,0 +1,15 @@
+
+
+
+
+import 'package:test/test.dart';
+
+void main() {
+  group('National ID Validation', () {
+    // test('Empty entry', () {
+    //   var nationalID = '';
+    //   var result = validateNationalID(nationalID);
+    //   expect(result, isA<Invalid>());
+    // });
+  });
+}